[Rename] LibraryController -> BooksRestController
